chore: fix new rust-1.94.0 clippy lints
Sweep new lints triggered by the toolchain bump: redundant closures
(`.map_err(|e| f(e))` -> `.map_err(f)`), `while let Some = iter.next()`
-> `for`, `iter().cloned().collect()` -> `to_vec()`, useless `into_iter()`
chain, useless u64 cast on FileMeta::size, filter_map(.then) -> filter().map(),
and a complex inner Pin<Box<dyn ...>> type alias to clean up file_listing.

+++ delta-kernel-datafusion-engine/src/compile/scan.rs
@@ -62,7 +62,7 @@ fn file_meta_to_partitioned(file: &FileMeta) -> Result<PartitionedFile, DeltaErr
     } else {
         StorePath::from(file.location.path())
     };
-    let mut pf = PartitionedFile::new(location, file.size as u64);
+    let mut pf = PartitionedFile::new(location, file.size);
     pf.object_meta.last_modified = Utc
         .timestamp_millis_opt(file.last_modified)
         .single()
